backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-pptx==0.6.23
pypdf==4.3.1
python-docx==1.1.2
